circle_mcp_server/pyproject.toml
[project]
name = "mcp-circle-developer-controlled-wallets"
version = "0.1.0"
description = "Circle Developer Controlled Wallet MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "mcp>=0.9.1",
    "requests",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]
